Move split function functional.split
