+++ corebehrt/functional/split.py
@@ -55,8 +55,19 @@ def split_pids_into_train_val(
 def load_train_val_split(
     data: PatientDataset, split_path: str
 ) -> Tuple[PatientDataset, PatientDataset]:
-    """
-    Load the train/val split from the given split path and return the corresponding data.
+    """Load predefined train/validation split from disk and filter data accordingly.
+
+    Args:
+        data: PatientDataset containing all patients
+        split_path: Path to directory containing train_pids.pt and val_pids.pt files
+
+    Returns:
+        Tuple containing:
+            - train_dataset: PatientDataset filtered to only include training patients
+            - val_dataset: PatientDataset filtered to only include validation patients
+
+    The split files should be PyTorch tensors containing patient IDs for each split.
+    The function expects files named 'train_pids.pt' and 'val_pids.pt' in the split_path directory.
     """
     splits = ["train", "val"]
     pids = load_predefined_splits(split_path, splits)
@@ -83,3 +94,19 @@ def get_n_splits_cv_pids(n_splits: int, train_val_pids: List[str]):
         train_pids = train_val_pids[train_idx].tolist()
         val_pids = train_val_pids[val_idx].tolist()
         yield train_pids, val_pids
+
+
+def split_into_test_and_train_val_pids(pids: list, test_split: float):
+    """Split patient IDs into test and train/validation sets.
+
+    Args:
+        pids: List of patient IDs to split
+        test_split: Fraction of patients to use for test set (between 0 and 1)
+
+    Returns:
+        Tuple of (test_pids, train_val_pids) containing the split patient IDs
+    """
+    test_pids = np.random.choice(pids, size=int(len(pids) * test_split), replace=False)
+    set_test_pids = set(test_pids)
+    train_val_pids = [pid for pid in pids if pid not in set_test_pids]
+    return test_pids, train_val_pids
